Extract test env. for `ProcessManager`
