+++ server/src/test/java/io/spine/server/procman/given/ProcessManagerTestEnv.java
@@ -1,0 +1,157 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.procman.given;
+
+import com.google.protobuf.Any;
+import com.google.protobuf.Empty;
+import com.google.protobuf.Message;
+import io.spine.core.CommandClass;
+import io.spine.core.CommandContext;
+import io.spine.core.CommandEnvelope;
+import io.spine.core.EventContext;
+import io.spine.core.Subscribe;
+import io.spine.protobuf.AnyPacker;
+import io.spine.server.command.Assign;
+import io.spine.server.commandbus.CommandBus;
+import io.spine.server.commandbus.CommandDispatcher;
+import io.spine.server.procman.CommandRouted;
+import io.spine.server.procman.ProcessManager;
+import io.spine.test.procman.ProjectId;
+import io.spine.test.procman.command.PmAddTask;
+import io.spine.test.procman.command.PmCreateProject;
+import io.spine.test.procman.command.PmStartProject;
+import io.spine.test.procman.event.PmProjectCreated;
+import io.spine.test.procman.event.PmProjectStarted;
+import io.spine.test.procman.event.PmTaskAdded;
+import io.spine.testdata.Sample;
+import io.spine.validate.AnyVBuilder;
+
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Set;
+
+import static io.spine.util.Exceptions.illegalStateWithCauseOf;
+
+/**
+ * @author Alexander Yevsyukov
+ */
+public class ProcessManagerTestEnv {
+
+    /** Prevents instantiation on this utility class. */
+    private ProcessManagerTestEnv() {}
+
+    @SuppressWarnings("UnusedParameters") // OK for test class.
+    public static class TestProcessManager extends ProcessManager<ProjectId,
+                                                                   Any,
+                                                                   AnyVBuilder> {
+
+        public TestProcessManager(ProjectId id) {
+            super(id);
+        }
+
+        /**
+         * Injects the passed CommandBus instance via Reflection since
+         * {@link #setCommandBus(CommandBus)} is package-private.
+         */
+        public void injectCommandBus(CommandBus commandBus) {
+            try {
+                final Method method = ProcessManager.class.getDeclaredMethod("setCommandBus",
+                                                                             CommandBus.class);
+                method.setAccessible(true);
+                method.invoke(this, commandBus);
+            } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
+                throw illegalStateWithCauseOf(e);
+            }
+        }
+
+        @Subscribe
+        public void on(PmProjectCreated event, EventContext ignored) {
+            getBuilder().mergeFrom(AnyPacker.pack(event));
+        }
+
+        @Subscribe
+        public void on(PmTaskAdded event, EventContext ignored) {
+            getBuilder().mergeFrom(AnyPacker.pack(event));
+        }
+
+        @Subscribe
+        public void on(PmProjectStarted event, EventContext ignored) {
+            getBuilder().mergeFrom(AnyPacker.pack(event));
+        }
+
+        @Assign
+        PmProjectCreated handle(PmCreateProject command, CommandContext ignored) {
+            getBuilder().mergeFrom(AnyPacker.pack(command));
+            return ((PmProjectCreated.Builder) Sample.builderForType(PmProjectCreated.class))
+                    .setProjectId(command.getProjectId())
+                    .build();
+        }
+
+        @Assign
+        PmTaskAdded handle(PmAddTask command, CommandContext ignored) {
+            getBuilder().mergeFrom(AnyPacker.pack(command));
+            return ((PmTaskAdded.Builder) Sample.builderForType(PmTaskAdded.class))
+                    .setProjectId(command.getProjectId())
+                    .build();
+        }
+
+        @Assign
+        CommandRouted handle(PmStartProject command, CommandContext context) {
+            getBuilder().mergeFrom(AnyPacker.pack(command));
+
+            final Message addTask = ((PmAddTask.Builder) Sample.builderForType(PmAddTask.class))
+                    .setProjectId(command.getProjectId())
+                    .build();
+            final CommandRouted route = newRouterFor(command, context)
+                    .add(addTask)
+                    .routeAll();
+            return route;
+        }
+    }
+
+    public static class AddTaskDispatcher implements CommandDispatcher<Message> {
+
+        private final List<CommandEnvelope> commands = new LinkedList<>();
+
+        @Override
+        public Set<CommandClass> getMessageClasses() {
+            return CommandClass.setOf(PmAddTask.class);
+        }
+
+        @Override
+        public Message dispatch(CommandEnvelope envelope) {
+            commands.add(envelope);
+            return Empty.getDefaultInstance();
+        }
+
+        @Override
+        public void onError(CommandEnvelope envelope, RuntimeException exception) {
+            // Do nothing.
+        }
+
+        @SuppressWarnings("ReturnOfCollectionOrArrayField") // OK for tests.
+        public List<CommandEnvelope> getCommands() {
+            return commands;
+        }
+    }
+}
